Feat | Show기능 분리
