+++ Assets/02.Scripts/02-8. UI/DailyReport/UI_ShowDailyReport.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 3836697c414cae345946a9ed05461d23
