+++ Assets/02.Scripts/02-8. UI/ProcessUI/OneCycleStartAndEnd.cs
@@ -8,7 +8,7 @@ public class OneCycleStartAndEnd : MonoBehaviour
     [SerializeField]
     private DailyReportModel _dailyReportModel;
     [SerializeField]
-    private UI_DailyReport _dailyReportUI;
+    private UI_ShowDailyReport _showDailyReport;
     
     [SerializeField]
     private TaxPayment _taxPayment;
@@ -73,7 +73,7 @@ public class OneCycleStartAndEnd : MonoBehaviour
             ? $"<color=#9A5195>**{_taxPayment.Tax}만큼 세금으로 나갔습니다**</color>"
             : $"<color=#9A5195>**세금 납부일까지 {_taxPayment.GetPaymentTerm(currentDate)}일 남았습니다.\n 납부할 세금 {_taxPayment.Tax}**</color>";
         _dailyReportModel.AddSpecialCommentText(taxMessage);
-        _dailyReportUI.ShowDailyReport();
+        _showDailyReport.ShowDailyReport();
         // UIManager.Instance.ReportUI.SpecialCommentText(taxMessage);
         // // 레포트 보여주고
         // UIManager.Instance.ReportUI.ShowReportUI();
