+++ Assets/03.Prefabs/UI/Prefab_DailyReport.prefab
@@ -1714,6 +1714,7 @@ GameObject:
   - component: {fileID: 6987657688561793243}
   - component: {fileID: 3119132307902995951}
   - component: {fileID: 5583716724902737875}
+  - component: {fileID: 1379699610912902734}
   m_Layer: 0
   m_Name: Prefab_DailyReport
   m_TagString: Untagged
@@ -1843,8 +1844,20 @@ MonoBehaviour:
   _fameChangeText: {fileID: 7231891759083617005}
   _questResultText: {fileID: 757293211443125485}
   _specialCommentText: {fileID: 1843530219157487688}
+--- !u!114 &1379699610912902734
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7572345628169123525}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 3836697c414cae345946a9ed05461d23, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
   _closeButton: {fileID: 5478682248965691327}
-  _hidePosition: {x: 0, y: 1200, z: 0}
+  _hidePosition: {x: 0, y: 100, z: 0}
   _onReportAudioClip: {fileID: 8300000, guid: c12eb330c021db547883e9d873964d1b, type: 3}
 --- !u!1 &7640215792143333506
 GameObject:
